Change layout to use Dropdowns.
With the number of desafios increasing, we need a better way to see them
on the screen. Remove the default output and replace it by a dropdown
containing the user name. Also, lots of formatting improvements (removed
the table, made things a bit tighter, etc).

Note: Attempted to put six pictures on a row, but that created problems
with devices having 992 pixel widths, since we use col-md-* for rows and
github usernames can get creatively long. :)